Move LineChart to components
