+++ src/components/LineChart/helpers.test.ts
