+++ src/components/LineChart/index.tsx
@@ -1,0 +1,3 @@
+import LineChart from './LineChart-container';
+
+export default LineChart;
